rename: presentation 모듈명을 server로 변경 #148
