+++ csalgo-server/src/main/java/kr/co/csalgo/server/scheduler/QuestionScheduler.java
@@ -1,0 +1,20 @@
+package kr.co.csalgo.server.scheduler;
+
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
+
+import kr.co.csalgo.application.problem.usecase.SendDailyQuestionMailUseCase;
+import lombok.RequiredArgsConstructor;
+
+@Component
+@RequiredArgsConstructor
+public class QuestionScheduler {
+
+	private final SendDailyQuestionMailUseCase sendDailyQuestionMailUseCase;
+
+	@Scheduled(cron = "0 0 8 * * MON-FRI", zone = "Asia/Seoul")
+	public void run() {
+		sendDailyQuestionMailUseCase.execute();
+	}
+}
+
