+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'csalgo'
-include "csalgo-common", "csalgo-infrastructure", "csalgo-domain", "csalgo-application", "csalgo-presentation"
+include "csalgo-common", "csalgo-infrastructure", "csalgo-domain", "csalgo-application", "csalgo-server"
